Abstract-tool: detects image window changes
Made a new Roi class to handle the arrow shape.
As side effect, it provides a way to store the arrow data within the
Roi, and to retrieve it when the user goes back to an image where there
is already an arrow drawn.
Something is still weird with the interaction status, tough.

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>