prepare-wiki-screenshot: rely on adjustView(), set default label
Recent developments in VIB make adjustView() really reliable, so let's
use that instead of some ad-hoc calculations that did not work in the
first place.

Also, default to the label "Fiji Is Just ImageJ".

The future plans are to integrate this into some developers-only
script that can be called with Fiji>New>Fiji Wiki Snapshot and which
lets you upload the snapshot to a special page for a (to-be-written)
RandomImage extension to be found, which places a random Fiji
snapshot on the front page, with a link to a Wiki page describing
the corresponding plugin.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>