Tutorial Maker: add a mode to edit Fiji News
As the Tutorial Maker brings most of what you need to edit news articles,
it was easy to teach it to do that.

Of course, the class was renamed to Wiki_Editor in the process to reflect
the change.

And now, this developer will never forget to add the News category anymore,
and he will neither pick the wrong date for the title (or forget the date
altogether)...

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>