Add an 'Auto Crop' plugin
The new Image>Adjust>Auto Crop plugin will try to cut as much border (in
the current background color) as possible.  It is possible to limit the
boundaries in which to crop with an initial rectangular selection which
is then refined.

For convenience, you can get the bounding box as a selection with
Edit>Selection>Select Bounding Box, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>