Jython interpreter: do not try to import anything in Java WebStart
Using Java WebStart, we cannot really just walk the classpath and
access the .jar files to see what's in them. So for the moment, we
cannot claim that we imported everything in that case.

Of course, this is just the official excuse to update
Jython_Interpreter.jar so it depends on the new Jython 2.5 rather
than the old 2.2.1 one.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>