Add the submodule "commons-math"
Commons Math has a variety of wonderfully useful stuff in it, and it was
wished for by Jean-Yves Tinevez a long time ago. Due to crooked eye, this
developer saw a .jar file size of more than 25 megabyte and refused to
include commons-math.jar into fiji.git, but in reality, the .jar weighs
in with a mere 800 kilobyte, which is fully worth it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>