Start to handle exceptions in the Script Editor
This does not do anything useful yet, except intercepting exceptions and
outputting them to stderr instead of a text window. Of course, the next
step will be to open the corresponding lines in the Script Editor.

This is a bit tricky, as the Java .class files only record what the base
name of the source file is, not the full path. For the moment, we just
look in the TextEditor for the base name.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>