Fiji (Linux): install an Alt key handler for Fiji's main window
It is not very intuitive to open menus with the F10 key (this developer had
to read the source of OpenJDK to realize that this is possible). The common
way is to press the Alt key and release it right away.

So let's support that.

For the moment, this is supported only on Linux (actually, only when the
current MenuBar's peer is an X11 one).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>