Add a FileDialog Decorator to add useful key handling to the dialog
Whenever this developer opens an image, he is annoyed that it is not
possible to jump to certain files by typing in their first characters.

This hack allows to do that.

It is still limited, though:
- only digits and letters (upper-case and lower-case are distinct) are
  handled
- the behavior for repeated characters is not as everywhere else yet.
- it does not make the dialog a drop target

Note that  this affects only platforms using X Windows, i.e. for the
moment only Linux. Both Windows and MacOSX use native file dialogs
which have proper key handling already.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>