Fiji Updater: fix long-standing bug when editing plugin details
Dependencies need to be editable by the developer, for example when the
crude logic detects a wrong dependency (this can happen when a class is
available in multiple .jar files, or when a platform-dependent system
class has dummy implementations; think Apple's com.eawt.* classes).

There was a bug when the tag was recorded incorrectly in the case more
than one dependency (or for that matter, category) was present.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>