Add a script to monitor Wikis
This script can be put into a cronjob to monitor Wikis based on the
MediaWiki engine.

It downloads the page Special:RecentChanges (logging in, if it found
the credentials in $HOME/.netrc), marks it up in a email-friendly way,
and writes out all lines it did not see the last time it ran.

The cached recent changes are stored in the file .recent-changes.<host>
in the current working directory.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>