Add a script to identify the commit corresponding to a .jar file
When you have a .jar file and do not know which commit actually introduced
the latest change on the source before releasing the plugin, you can now
run this script. If you already know the timestamp, or the checksum,
you can use it to determine said commit.

Suggested by Mark Longair.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>