Add a Python script for configuring submodules.
The script determines whether the submodule has already been initialized,
initializes it if necessary, switches to the appropriate branch if needed,
and pulls the latest changes from origin.

The script also similarly initializes any prerequisite submodules,
as defined in Fiji's main Fakefile.

Currently, the script uses os.system a lot with some sh-isms
(specifically && and 2>&1), so no chance of working on Windows.