work in progress using java Calendar, Date and SimpleDateFormat, and preview with slider
Preview with slider in the GUI almost works now,
but there is a problem with the timing of the calles under the hood,
meaning that timestamps are drawn and not removed when the slider is moved.

the method digitalString now is passed (double time)
and uses a GregorianCalendar object to hold the time in milliseconds
which is calculatede according to the time multiplied
by a factor determined by the unit of the time selected in the GUI.
Choosing a suffix/unit other than those used by Date object
should result in a IJ.error message.
its a bit ugly you know... the way the years and months are calculated,
since months are differnt lengths and stuff....