Add an Exact Signed Euclidean Distance Transform
Based on the exact Euclidean distance transform, add a modified plugin
that calculates the signed transform, i.e. for background pixels, the
negative distance to the nearest foreground pixel.

This implementation makes one simplification for ease of computation:
for diagonal foreground/background neighbors, the distance 0.5 is
assumed instead of the correct 0.5*sqrt(2) and 0.5*sqrt(3).

This plugin was requested by Erwin Frise.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>