IEP2: testing the new logic on a single operator
This starts the development of Image Expression Parser 2.0.

As requested by Johannes Schindelin, it would be nice to have the
possibility for the IEP to include functions such as gaussian filter or
convolution and so on. These methods are already there in the imglib
library, one by one.

However, as in its current state, the IEP is incapable of dealing with
such functions: it simply applies the expression to each pixel
individually.

A solution would be to make to parser aware of Image<T> object as a
whole, and not limiting it to work on individual pixel value.

This implies rewriting a major part of the parser logic, currently
implementd in JEP to work on scalars only (but see djep for more).

We started this with this commit, implementing the multiply operator to
work on imglib Image<T> object as a whole. Right now this operator does
element by element multiplication.

Surprisingly, it works.

As of now, the results are always cast to FloatType, to avoid overflow
and underflow, which prevents us to use the Type.mult() method
unfortunately. A better solution might arise.

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>