FijiClassLoader: offer retrotranslation of the offending .jar file
Unfortunately, we cannot try to load the class again, because of a
catch-22: To refresh the menus an make use of the new .jar file, we
would have to replace the FijiClassLoader by a new instance thereof,
release the old FijiClassLoader (to release the .jar files), and return
within the same stackframe. This is unfortunately not possible.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>