Add a method to get a generic component path of a Component
Use the static method fiji.Main.getPath(Component) to get a unique path
(actually, the dialog is not checked for uniqueness) to a component.

This path looks something like this:

	Some Dialog>Panel[2]>TextField{Name:}

meaning the first text field after the label "Name:" in the 2nd Panel
of the dialog with the title "Some Dialog".

The function fiji.Main.getComponent(String path) is the companion to
getPath(Component component): use the former to retrieve the component
using the result of the latter.

This contains fixes by Andreas Wiese.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>