FijiClassLoader: if the property jnlp_class_map is set, try to load it
The property will be interpreted as a URL to a .jar file that contains
a file called 'class.map'. The lines of said file are interpreted as
class -> .jar file names, where the .jar file names are expected to
be relative to the directory of the first .jar file.

Example:

The class.map file in http://blah.org/some/where/blub.jar could look
like this:

-- snip --
my.package.Class1 incrediboy.jar
my.package.Class2 bomb-voyage.jar
-- snap --

If the class "my.package.Class1" wants to be loaded and is not found in
the classpath, the classpath is amended by
http://blah.org/some/where/incrediboy.jar and the class gets another go
at being loaded.

This is very useful if we want to start Fiji via WebStart and load only
those classes that are required for the user-specified actions, rather
than loading >40m of .jar files when only a 2k .jar file would have been
sufficient.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>