Series labeler - new methods handling meta data units
A method for selecting custom unit format from stack stack selected
also a method for bringing the meta data unit got from imp to the gui.

When there is meta data in imp.Calibration with units that we can use
to fill the GUI fields and choosing the right default
value for drop down menus, we now do that.

If the user selects something else or puts in their own custom suffix
then we use those not the data read from the meta data in the imp.