IEP2: changed the event mechanism
Following advices from Albert Cardona, Johannes Schindelin and Johan
Henriksson, rewrote the way the gui and the plugin communicate.

Now the run() method of the plugin simply launches the GUI, which handle
all events, and delegate calculations back to the plugin.

The main consequence of these advices is that now, the plugin works.

Also, for future portability, the plugin now implements the
imglib.algorithm.OutputAlgorithm interface.
Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>