Add clojure-contrib as a submodule again
It is absolutely not in the spirit of Open Source to take away the easy
way to fiddle with source code and recompile it. Binary blobs have hurt
us already much, and there is not at all a need to repeat those mistakes.

There are other bad consequences of just providing a binary blob, such
as the outright impossibility to know which version the blob is based
on, and whether somebody has been creative with the source while
producing the binary blob.

Therefore, this commit reverts "Updated clojure-contrib.jar, which is no
longer a submodule.", which did not really remove the submodule
completely to begin with.

This commit does _not_ re-add the precompiled/ .jar, as the commit
it reverts already cluttered the repository with useless megabytes.

Instead, we now rely on the previously decided-upon method to download
the .jar file from the updater.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>