Update clojure-contrib to the current 'master'
This update requires the helper script bin/maven.sh, which was introduced
just for this purpose in a previous commit.

Most likely, the fans of Clojure need to identify a stable version of
clojure-contrib and update the submodule again. At least, now we do not
clutter the repository with precompiled .jar files anymore.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>