Removing clojure-contib.jar from fiji.
The clojure-contrib.jar contains the clojure.contrib.* libraries,
which change very rapidly and are not suitable for distribution
as stable binaries (i.e. they are not stable).

In addition, the main features of clojure.contrib.* that motivated
its inclusion in fiji have now been transferred to clojure.jar,
as of version 1.2 of the latter (and which has been committed to fiji).

Removing a submodule needs:
 * remove entry from .gitmodules
 * remove entry from .git/config
 * remove staged-plugins/clojure-contrib.Fakefile
 * remove all entries in the fiji Fakefile

I tried as well "git submodule remove clojure-contrib"
and "git rm jars/clojure-contrib.jar"
and "git rm precompiled/clojure-contrib.jar"
but all of them were claimed not to be in git.