log-plugin-commits: add [-v [-v]] flags to show javap differences
Sometimes, the only difference lies in a single class, and it is useful
to look at the differences of the byte code.

In even rarer cases, the difference of two .class files is only in the
(invisible) annotations. You need to pass the '-v' option twice to get
that level of detail.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>