Coloc 2: added class to calculate a 2D intensity histogram
aka. scatterplot or cytofluorogram etc, the 2D histogram for an image pair
gives us an idea of the correlation of 2 images over space.
It will be used as input for other methods, and for display by itself.

TODO - remove the nasty if (count < 65535) limitation as its plain wrong.
This is a relice from the colocalisation threshold plugin

TODO - need to really store the whole full resolution 2D histogram
with the native intensity scales, not rescaled to 0-255.
For display, the full/native scale 2D histogram image can then be
resized and intensity scaled as required.