Created two separate findMaxima pipelines - one for 2D images, and one for 3D. Also did some code tidying/commenting.
I had hoped that not performing if/else statements to check whether the image was 2D or 3D in the while loop would improve performance, but it didn't really help. Currently, there are 2D and 3D versions of many functions, which look identical aside from the way a few variables are declared.