Made it so each dimension is scaled separately.
Before, all the dimensions were scaled equally to make blobs have a diameter of 10 pixels. However, we don't always want to scale each dimension by the same amount (for example, if 30x30 in x and y, may not be 30 in z, so scale z separately to make the blob as spherical as possible).