Simple_Neurite_Tracer: just warn about out-of-order fill IDs
The bug fixed in c4b6f91bf68eddb6fe0e may mean that people have
saved out files which can no longer be loaded due to all the
fills having id="0", but which are otherwise fine.  Deal with
this by downgrading the error on out-of-order fill IDs to just a
message to the log.