Embryo_Tracker: Moved the actually thresholding step to the 3D viewer section to better control what's displayed. Improved how the histogram for thresholding is created. Made it so threshold returned is a pixel intensity and not a bin. BlobContrastScorer: Added out of bounds strategy (mirror) for the ROI cursor. BlobVarianceScorer: Added out of bounds strategy (mirror) for the ROI cursor. BlobBrightnessScorer: Fixed the usage of the ROI cursor to be a square completely inside the sphere that should be the blob (inscribed cube).
Embryo_Tracker now creates histogram using sqrt(N) approach to determine the number of bins. Still using Otsu's method for the actual thresholding. Also, now an ArrayList of the extrema for each frame as well as an ArrayList of the thresholds for each frame are returned from the exec method, and used by the 3D viewer to display the correct thresholds.

Signed-off-by: Nick Perry <nicholas.m.perry@gmail.com>