Change the implementation of ij3d.Pipe.makeTube
The previous implementation of ij3d.Pipe.makeTube failed for
sequences of points where any two adjacent points differed only
in z.  In addition, the old method produce strange results when
the change in y between points changed between positive and
negative.  This new implementation works differently - it tries
to match up the edge points of each disc such that if the two
adjacent disks were overlaid, the ends of the radius running
along their line of intersection would be matched up.

This also slightly changes how the tangent to the line at each
point is calculated.  Previously, the tangent at point i was
considered to be the vector from i to i+1, whereas now (when a
point on either side of i exists) it will be the vector
from i-1 to i+1

In my measurements the performance is slightly improved, but that
may just be due to precalculating the sin() and cos() values.