Eclipse projects generator: use default JVM library
In a vain attempt to relieve the developer who wants something that
Just Works and Without Any Hassels, I tried to force JVM-1.6: a few
projects in Fiji use the @Override directive which Fiji's javac handles
just fine, but not Eclipse's if it targets Java 1.5.

The problem is that Eclipse does not respect that setting on MacOSX, at
least Eclipse version 3.5, not even when you set the default JVM in
Eclipse to 1.6. So until Eclipse learns some sanity in that respect, the
developer will have to click Properties>Java Build Path>Libraries,
double click the JVM entry and select 1.6 explicitely. Funnily enough,
Eclipse respects _that_ (but the developer has to do that even if the
workspace default was _already_ 1.6).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>