Spot: Added a method for directly returning the value associated with a given feature. Embryo_Tracker: Now uses the new method in the Spot class for getting values for a specific feature.
Before, in Embryo_Tracker.java, whenever a Feature's value was needed, the logic was to use getFeatures() to return the entire Feature->Value map, and use the map's .get() method to retrieve the feature. Now, with the Spot's new method, a getFeature(Feature feature) method exists, so that a user can request the value for a specific feature instead of having to work with the whole Feature map.

Signed-off-by: Nick Perry <nicholas.m.perry@gmail.com>