Embryo_Tracker: working (hf) BlobMorphology
... based on inertia moments.
We assume we get an ellipsoid, with centroid given as spot center, and
we compute the inertia moments for axes x, y and z.
We put them in the moment of intertia tensor, which we diagonalize,
using Jama. The eignvalues are used to retrieve the ellipsoid semiaxes
length, and the eigenvectors are used to get the ellipsoid orientation.

A test main() method is provided, however I found myself unable to
generate ellipsoid with main axis not belonging to the XY plane.
Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>