NearestNeighborLinker: Now returns a HashMap as the result, instead of ArrayList. ObjectTracker: modified to use a HashMap now returned from the NearestNeighborLinker class as the result.
Previously, NearestNeighborLinker returned an ArrayList< ArrayList<Spot> > as the result, which didn't keep the association between a Spot in time t, and it's potentials in t+1. Now, a HashMap is returned, Spot in t --> Spots in t+1 to make this relationship explicit. So now getResult() returns HashMap< Spot, ArrayList<Spot> >.

Signed-off-by: Nick Perry <nicholas.m.perry@gmail.com>