Update ImageJA to interpret the IJ_PREFS_DIR variable
As suggested by Stephan Saalfeld in bug 187, the IJ_PREFS_DIR
environment variable is now inspected before constructing a default path
to the directory where IJ_Prefs.txt lives (along with the histories of
the scripting interpreters).

This is necessary to be able to override the setting on Windows, in case
the user directory obtained via System.getProperty("user.home") is not
writable or even invalid.

Noticed by Christian Goosmann.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>