Emryo_Tracker: finished implementing LAPTracker, but buggy
LAPTracker, and the two cost matrix classes used by it, are *written* but seemingly non-functional. Both cost matrices are created, and the LAP is executed on them. The results are compiled into track segments and final tracks, respectively. But the LAP for gap closing / merging / splitting doesn't finish. So something is wrong with either the Hungarian Algorithm implementation, or my cost matrix for that step.

Signed-off-by: Nick Perry <nicholas.m.perry@gmail.com>