Remove JNA from Jython and JRuby
Especially JRuby's copy is incredibly outdated. We have an up-to-date
jna.jar in jars/ which was downloaded from the official JNA website
(not built from scratch, because jna.jar contains support for many
more platforms than Fiji, and there is still a very loose plan to have
a minimal shell script for unsupported operating systems/CPUs with
support for Java >= 1.5).

The trick was performed by

	for f in jars/jruby.jar precompiled/jython.jar
	do
		zip -d $f 'com/sun/jna/*'
	done

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>