Add a script to compile JNI libraries
This needs to be handled by Fiji Build in the future, also to handle
cross-compilation correctly.

The idea is to have libraries in lib/<platform>/lib<jarname>.<ext>
generated when CompileJar detects that there were some .c files in the
sources.

Another variable would list the required libraries, and yet another
variable would trigger building for all platforms using the cross
compilers (failing if they are not available).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>