Updater: store timestamp of update sites in db.xml.gz
This change is rather involved, since we can no longer just download the
db.xml.gz file, but need to write a different file locally.

While at it, prepare the database downloading and the up-to-date check for
multiple update sites.

For the moment, the uploading part is not prepared for multiple update
sites, but continues to have hard-coded support for the principal Fiji
site.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>