Coloc 2: Add random image generation macro
This macro uses the Process>Math>Macro command to create a synthetic
image. The formula that generates the image comes from MathMacroDemo
macro from imageJ example macros page
http://rsbweb.nih.gov/ij/macros/examples/MathMacroDemo.txt The image
is a sin wave in x and a cos wave in y, giving a regular grid of blobs
that look a bit like diffraction limited images of sub resolution
objects.