Bug_Submitter: also include JAVA_HOME and fiji.dir in the output
This is in preparation for the proposed change to remove the
possibility to override the chosen JVM with the JAVA_HOME
environment variable.  See the discussion here:

  http://groups.google.com/group/fiji-devel/browse_thread/thread/29d98d6d1e5bad8/255a0b325c12b413

This should help us to figure out problems that might occur after
this switch where people who are (perhaps inadvertantly) relying on
the JAVA_HOME override report bugs.