Refresh Javas: re-add convenient compile & run of Fakefile targets
When compiling a class that is part of a .jar file, but does not
implement the PlugIn nor the PlugInFilter interface, and does not have a
static main() method, either, we used to fall back to offering to start
a plugin defined in the .jar file's plugins.config.

Make that work again.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>