Fiji Build: leave variables overridden from the environment as-are
We introduced the ENVOVERRIDES(<name>) construct to handle JAVA_HOME.
But we tried to be clever and stripped the Fiji root from the
overridden value because we used to use that setting in the Fiji
launcher.

That special treatment was misguided for two reasons:

1) the Java path hardcoded into the Fiji launcher should not be
   overridden by the environment variable JAVA_HOME (which is why
   we hardcode the value of FIJI_JAVA_HOME in the Fakefile now)
2) nobody says that ENVOVERRIDES() will be used for paths always

So just don't process the value obtained from the environment variable.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>