Compile the unexpanded Java path into the Fiji launcher
When compiling from another working directory than the Fiji root, we
need to prefix the Java path to be able to get at, say, tools.jar.

But the Fiji launcher needs to know the path relative to the Fiji root,
so let's use the unexpanded (unprefixed) Java path in the launcher.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>