MenuItemDiverter: reset when windows focus was gained again
When things go wrong with the menu item diverter, they go wrong in a big
way. It is usually not even possible to reset actions with the Escape key.
So let's be nice and assume that switching to another window and back means
that we do not want to divert the menu items any longer.

We don't reset when the focus is lost because one might want to start the
diverted action via the Command Launcher.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>